Refactor 7 - Refactor utilities (#1030)
Summary:
Pull Request resolved: https://github.com/facebookresearch/spdl/pull/1030

Part 7 of https://github.com/facebookresearch/spdl/pull/1026

Split `_utils` into `_common` and user-facing utilities.

Differential Revision: D84881510
