+++ src/spdl/pipeline/_components/_hook.py
@@ -15,7 +15,7 @@ from contextlib import asynccontextmanager, AsyncExitStack
 from dataclasses import dataclass
 from typing import AsyncContextManager, TypeVar
 
-from .._utils import create_task
+from .._common._misc import create_task
 from ._common import _periodic_dispatch, _StatsCounter, _time_str
 
 __all__ = [
