+++ src/spdl/pipeline/_pipeline.py
@@ -21,7 +21,7 @@ from typing import Any, Generic, TypeVar
 
 from spdl._internal import log_api_usage_once
 
-from ._utils import create_task
+from ._common._misc import create_task
 
 __all__ = ["Pipeline"]
